♻ extract animateRoundTransition
